requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
gunicorn
Werkzeug==3.0.0
WTForms==3.0.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Bootstrap-Flask==2.2.0
requests==2.25.1
